requirements.txt
langchain-chroma==0.1.4
langchain-core==0.3.51
langchain-openai==0.2.6
langchain-text-splitters==0.3.2
python-dotenv==1.0.1
langchain-ollama==0.3.1
faiss-cpu==1.10.0
beautifulsoup4>=4.12.0
html2text==2024.2.26